introduce agents module
