+++ playpen/agents/__init__.py
@@ -1,0 +1,8 @@
+from playpen.agents.base import BaseAgent
+from playpen.agents.clem import ClemAgent, ClemObservation
+
+__all__ = [
+    "BaseAgent",
+    "ClemAgent",
+    "ClemObservation"
+]
